refactor: vertical slice architecture, schema versioning, lint clean-up
  Task-type vertical slice (src/task-types/)
  - Co-located types, TaskView, Copier, and sub-components under
    src/task-types/{qa,generation,rag}/
  - Renamed task types to current taxonomy: single-turn rag→qa,
    text_generation/json_generation→generation, chat→rag
  - Added taskTypeRegistry in src/task-types/index.ts; dispatchers
    in Task.tsx and TaskCopier.tsx replaced with registry lookups
  - Split exportData out of processor.ts into src/exporter.ts

  Schema versioning (src/migrator.ts)
  - CURRENT_SCHEMA_VERSION=2 with versioned migration chain table
  - migrateV1toV2 deep-clones and renames task_type fields (snake_case)
  - Wired into dataloader.ts and DataUploader.tsx before camelCaseKeys;
    migrated flag threads through to Data.migrated
  - exporter.ts stamps schema_version on all exported files

  Lint clean-up (0 errors, 0 warnings)
  - Eliminated setState-in-effect/useMemo anti-patterns: derived
    ExpressionBuilder error, Filters visibility, DataVerification
    pagination, and PredictionsTable warning from render instead
  - Fixed Card.tsx dynamic heading tag (static HEADING_TAGS map)
  - Fixed SkeletonTask.tsx Math.random() in render
  - Added targeted eslint-disable comments with explanations for
    legitimate patterns (stable refs, intentional range resets)
  - Fixed loading.tsx copyright year
