+++ src/processor.ts
@@ -18,7 +18,6 @@
 
 import { isEmpty, isNumber } from 'lodash';
 import { hash } from '@/src/utilities/strings';
-import { snakeCaseKeys } from '@/src/utilities/objects';
 import {
   Data,
   MetricValue,
@@ -27,7 +26,6 @@ import {
   DisqualificationReason,
   DisqualifiedTasks,
   Task,
-  RetrievedDocument,
   Notification,
 } from '@/src/types';
 
@@ -99,6 +97,7 @@ function disqualifyEvaluation(
 
 export function processData(
   data: RawData,
+  migrated = false,
 ): [Data, DisqualifiedTasks, Notification[]] {
   const notifications: Notification[] = [];
 
@@ -293,134 +292,11 @@ export function processData(
       evaluations: qualifiedEvaluations,
       annotators: Array.from(annotators),
       numTasks: qualifiedTasks.length,
+      // Carry the migration flag so exportData can show a one-time toast
+      // when the researcher downloads the upgraded file.
+      ...(migrated && { migrated: true }),
     },
     disqualifiedTasks,
     notifications,
   ];
-}
-
-export function exportData(
-  data: Data | undefined,
-  tasks: Task[] | undefined,
-): boolean {
-  if (data) {
-    let dataToExport: RawData = {
-      name: data.name,
-      ...(data.exampleId && { exampleId: data.exampleId }),
-      ...(data.filters && { filters: data.filters }),
-      models: data.models,
-      metrics: data.metrics,
-      ...(data.documents && {
-        documents: data.documents,
-      }),
-      tasks: data.tasks,
-      evaluations: data.evaluations.map((evaluation) => {
-        return {
-          taskId: evaluation.taskId,
-          modelId: evaluation.modelId,
-          modelResponse: evaluation.modelResponse,
-          annotations: evaluation.annotations,
-          ...(evaluation.contexts && { contexts: evaluation.contexts }),
-        };
-      }),
-    };
-
-    if (tasks) {
-      // Ensure every task has a flagged property before export
-      tasks.forEach((task) => {
-        if (!task.hasOwnProperty('flagged')) {
-          task.flagged = false;
-        }
-      });
-
-      // Subset export: only include documents and evaluations for the given tasks
-      if (data.tasks.length !== tasks.length) {
-        const documentsMap: Map<string, RetrievedDocument> = new Map(
-          data.documents?.map((document) => [document.documentId, document]),
-        );
-
-        const relevantDocuments: Set<RetrievedDocument> =
-          new Set<RetrievedDocument>();
-        const relevantTaskIds: Set<string> = new Set<string>();
-
-        // Collect referenced document IDs from the subset of tasks
-        tasks.forEach((task) => {
-          relevantTaskIds.add(task.taskId);
-
-          if (documentsMap.size !== 0) {
-            task.contexts?.forEach((context) => {
-              if (typeof context !== 'string') {
-                const referenceDocument = documentsMap.get(context.documentId);
-                if (referenceDocument) {
-                  relevantDocuments.add(referenceDocument);
-                }
-              }
-            });
-          }
-        });
-
-        dataToExport = {
-          name: data.name,
-          ...(data.exampleId && { exampleId: data.exampleId }),
-          ...(data.filters && { filters: data.filters }),
-          models: data.models,
-          metrics: data.metrics,
-          ...(relevantDocuments.size !== 0 && {
-            documents: Array.from(relevantDocuments),
-          }),
-          tasks: tasks,
-          evaluations: data.evaluations
-            .filter((evaluation) => relevantTaskIds.has(evaluation.taskId))
-            .map((evaluation) => {
-              return {
-                taskId: evaluation.taskId,
-                modelId: evaluation.modelId,
-                modelResponse: evaluation.modelResponse,
-                annotations: evaluation.annotations,
-                ...(evaluation.contexts && { contexts: evaluation.contexts }),
-              };
-            }),
-        };
-      } else {
-        // Full export: all tasks provided, just copy tasks with their flagged state
-        dataToExport = {
-          name: data.name,
-          ...(data.exampleId && { exampleId: data.exampleId }),
-          ...(data.filters && { filters: data.filters }),
-          models: data.models,
-          metrics: data.metrics,
-          ...(data.documents && {
-            documents: data.documents,
-          }),
-          tasks: tasks,
-          evaluations: data.evaluations.map((evaluation) => {
-            return {
-              taskId: evaluation.taskId,
-              modelId: evaluation.modelId,
-              modelResponse: evaluation.modelResponse,
-              annotations: evaluation.annotations,
-              ...(evaluation.contexts && { contexts: evaluation.contexts }),
-            };
-          }),
-        };
-      }
-    }
-
-    // Trigger a browser download via a temporary anchor element
-    var element = document.createElement('a');
-    element.setAttribute(
-      'href',
-      'data:application/json;charset=utf-8, ' +
-        encodeURIComponent(JSON.stringify(snakeCaseKeys(dataToExport))),
-    );
-    element.setAttribute('download', 'analytics.json');
-
-    document.body.appendChild(element);
-    element.click();
-    document.body.removeChild(element);
-
-    return true;
-  }
-
-  return false;
 }
