+++ src/task-types/generation/TaskView.module.scss
