+++ src/task-types/qa/Copier.module.scss
@@ -16,8 +16,38 @@
  *
  **/
 
-import SkeletonExample from '@/src/views/example/SkeletonExample';
+@use '@carbon/react/scss/spacing' as *;
 
-export default async function Loading() {
-  return <SkeletonExample />;
+.container {
+  display: flex;
+  flex-direction: column;
+  row-gap: $spacing-02;
+}
+
+.copyFormatSelectors {
+  display: flex;
+  column-gap: $spacing-03;
+}
+
+.copyFormatSelectors > div > label {
+  background-color: var(--cds-background) !important;
+}
+
+.previewBox {
+  height: 75%;
+  background-color: var(--cds-background);
+}
+
+.heading {
+  padding-bottom: $spacing-03;
+}
+
+.warningContainer {
+  display: inline-flex;
+  align-items: center;
+  column-gap: $spacing-01;
+  font-size: 0.75rem;
+  font-weight: 400;
+  line-height: 1rem;
+  letter-spacing: 0.32px;
 }
