+++ src/task-types/rag/components/DocumentsViewer.tsx
@@ -26,7 +26,7 @@ import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 import rehypeRaw from 'rehype-raw';
 
-import { ToolMessageDocument } from '@/src/types';
+import { ToolMessageDocument } from '@/src/task-types/rag/types';
 
 import classes from './DocumentsViewer.module.scss';
