+++ src/views/home/Card.tsx
@@ -27,7 +27,14 @@ import classes from './Card.module.scss';
 
 type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;
 
-const getHeading = (level: HeadingLevel) => `h${level}` as const;
+const HEADING_TAGS = {
+  1: 'h1',
+  2: 'h2',
+  3: 'h3',
+  4: 'h4',
+  5: 'h5',
+  6: 'h6',
+} as const;
 
 export interface CardProps {
   icon: ComponentType<{ size?: number; className?: string }>;
@@ -55,7 +62,7 @@ function Card({
   disabled,
 }: CardProps) {
   const id = useId();
-  const Heading = getHeading(headingLevel);
+  const Heading = HEADING_TAGS[headingLevel];
   return (
     <section className={cx(classes.root, disabled && classes.disabled)}>
       {href && actionText && (
